update stage to processed
